+++ src/main/java/com/recodepro/enxametech/repository/AlunoRepository.java
@@ -3,5 +3,5 @@ package com.recodepro.enxametech.repository;
 import com.recodepro.enxametech.model.Aluno;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlunoRepository extends JpaRepository<Aluno, Integer>{
+public interface AlunoRepository extends JpaRepository<Aluno, Long>{
 }
